refactoring firebase utils and calendar page

+++ src/firebase-calendar-utils.tsx
@@ -13,17 +13,23 @@ export const stockCalendarData = async (
   try {
     const userRef = db.collection('user-data').doc(userId);
     const calendarRef = userRef.collection('calendar');
-    const calendarData = await calendarRef.limit(1).get();
 
+    // if calendar subcollection doesn't exists for this user
+    // create it with empty value
+    const calendarData = await calendarRef.limit(1).get();
     if (calendarData.empty) {
       calendarRef.doc(dateMeal).set({});
     }
 
     const dayRef = calendarRef.doc(dateMeal);
     const dayDoc = await dayRef.get();
-    const { id } = recipe;
-    const recipeIdString: string = id.toString();
 
+    const { id } = recipe;
+    const recipeIdString = id.toString();
+
+    // if doc for this date does not exists create it and
+    //        initialize the array with the first id
+    // if exists update the array adding the id
     if (!dayDoc.exists) {
       await dayRef.set({
         date,
@@ -36,18 +42,27 @@ export const stockCalendarData = async (
       });
     }
 
-    const recipeRef = db.collection('recipes').doc(recipeIdString);
-    const recipeDoc = await recipeRef.get();
-
-    if (!recipeDoc.exists) {
-      await recipeRef.set({
-        ...recipe,
-      });
-    }
+    // add the recipe datas to recipes collection
+    await addRecipeToRecipes(recipeIdString, recipe);
 
     return true;
   } catch (error) {
     return false;
+  }
+};
+
+export const addRecipeToRecipes = async (
+  recipeIdString: string,
+  recipe: Recipe
+) => {
+  const recipeRef = db.collection('recipes').doc(recipeIdString);
+  const recipeDoc = await recipeRef.get();
+
+  // add recipe only if doesn't exists
+  if (!recipeDoc.exists) {
+    await recipeRef.set({
+      ...recipe,
+    });
   }
 };
 
@@ -65,9 +80,20 @@ export const removeRecipeFromCalendar = async (
     recipes_list: firebase.firestore.FieldValue.arrayRemove(recipeId),
   });
 
-  const datas = dayDoc.data();
+  // if the id is the last one of the recipes_list array
+  // for this day
+  // remove the parent document (the day)
+  removeDayFromCalendar(dayDoc, dayRef);
+};
 
-  if (datas?.recipes_list.length === 1) {
-    dayRef.delete();
+export const removeDayFromCalendar = async (
+  dayDoc: firebase.firestore.DocumentData | undefined,
+  dayRef: firebase.firestore.DocumentReference<firebase.firestore.DocumentData>
+) => {
+  if (dayDoc) {
+    const datas = dayDoc.data();
+    if (datas?.recipes_list.length === 1) {
+      dayRef.delete();
+    }
   }
 };
